fix(doc): attempt fixing missing file watcher
.htm files of usability demonstration (docs/) are not part of dependency graph,
therefore not being watched by webpack, even though explicitly defined through a
globbing pattern. A workaround without having to include the files into the
dependency graph would be to define a custom middleware that sends a message to
the webpack-dev-server client to force reloading. Since chokidar is a dependency
of webpack-dev-server, it's reasonable to just use it for watching the files.
HOWEVER with my setup this doesn't work. Globbing is not working in chokidar in
an MSYS2/MinGW environment.

I filed a bug report for chokidar: https://github.com/paulmillr/chokidar/issues/1419

Added a TODO and will call it a day for now. During development, changes to the
.htm files will not work with hot reloading at the moment. But it has been that
way since the start of the project... So i'm used to it.